Refactor read helper into io module
