+++ wandas/io/__init__.py
@@ -1,4 +1,5 @@
+from .read import read
 from .wav_io import write_wav
 from .wdf_io import load, save
 
-__all__ = ["load", "save", "write_wav"]
+__all__ = ["load", "save", "write_wav", "read"]
